+++ tests/testthat/test-layout_heatmap.R
@@ -38,7 +38,7 @@ testthat::test_that("`ggheatmap` add `quad_switch()` works well", {
         "heatmap-hmanno-null-guides",
         ggheatmap(matrix(1:9, nrow = 3L)) +
             scale_fill_viridis_c() -
-            plot_action(guides = NULL) +
+            plot_align(guides = NULL) +
             quad_anno("r") +
             align_dendro()
     )
@@ -47,17 +47,17 @@ testthat::test_that("`ggheatmap` add `quad_switch()` works well", {
         ggheatmap(matrix(1:9, nrow = 3L)) +
             scale_fill_viridis_c() +
             quad_anno("l") -
-            plot_action(guides = "l") +
+            plot_align(guides = "l") +
             align_dendro(aes(color = branch))
     )
     expect_doppelganger(
         "heatmap-hmanno-align-action-guides",
         ggheatmap(matrix(1:9, nrow = 3L)) +
             scale_fill_viridis_c() -
-            plot_action(guides = NULL) +
+            plot_align(guides = NULL) +
             quad_anno("r") +
             align_dendro() +
-            plot_action(guides = "l")
+            plot_align(guides = "l")
     )
 })